Internal abstract class holding managed state
